refactor: use shorter client import paths where possible
