+++ pages/events/index.tsx
@@ -3,13 +3,13 @@ import { clamp } from "@foxkit/util/clamp";
 import { useStore } from "@nanostores/react";
 import { Searcher, sortKind } from "fast-fuzzy";
 import { api } from "~/client/api";
-import { EventList } from "~/client/components/EventList";
-import Headline from "~/client/components/Headline";
-import { Input } from "~/client/components/Input";
-import Meta from "~/client/components/Meta";
-import { Scroller } from "~/client/components/Scroller";
-import Section from "~/client/components/Section";
 import { settingsStore } from "~/client/stores/settingsStore";
+import { EventList } from "~/components/EventList";
+import Headline from "~/components/Headline";
+import { Input } from "~/components/Input";
+import Meta from "~/components/Meta";
+import { Scroller } from "~/components/Scroller";
+import Section from "~/components/Section";
 import type { EventsPageProps } from "~/pages/EventsPage/static";
 import styles from "~/pages/EventsPage/EventsPage.module.css";
