+++ pages/upgrades.tsx
@@ -4,19 +4,19 @@ import { useStore } from "@nanostores/react";
 import { Searcher } from "fast-fuzzy";
 import type { MatchData } from "fast-fuzzy";
 import { api } from "~/client/api";
-import { ActionButton } from "~/client/components/Button";
-import { CardGrid } from "~/client/components/Card";
-import Headline from "~/client/components/Headline";
-import { IconAtlas, IconHelp } from "~/client/components/icons";
-import { InlineSvg } from "~/client/components/InlineIcon";
-import { Input } from "~/client/components/Input";
-import Meta from "~/client/components/Meta";
-import { Modal, ModalMenu } from "~/client/components/Modal";
-import { NoSSR } from "~/client/components/NoSSR";
-import { Scroller } from "~/client/components/Scroller";
-import Section from "~/client/components/Section";
 import { DataContext } from "~/client/contexts";
 import { settingsStore } from "~/client/stores/settingsStore";
+import { ActionButton } from "~/components/Button";
+import { CardGrid } from "~/components/Card";
+import Headline from "~/components/Headline";
+import { IconAtlas, IconHelp } from "~/components/icons";
+import { InlineSvg } from "~/components/InlineIcon";
+import { Input } from "~/components/Input";
+import Meta from "~/components/Meta";
+import { Modal, ModalMenu } from "~/components/Modal";
+import { NoSSR } from "~/components/NoSSR";
+import { Scroller } from "~/components/Scroller";
+import Section from "~/components/Section";
 import { UpgradeCard } from "~/pages/UpgradesPage/components";
 import { FiltersForm } from "~/pages/UpgradesPage/components/FiltersForm";
 import {
