+++ src/client/components/Card/CardHero.tsx
@@ -1,4 +1,4 @@
-import { IconFace } from "~/client/components/BorderedIcon";
+import { IconFace } from "~/components/BorderedIcon";
 import styles from "./CardHero.module.css";
 
 type CardHeroProps = Pick<
