+++ src/client/components/CommandCodeIcon/CommandCodeIcon.tsx
@@ -1,7 +1,7 @@
 import ccat from "classcat";
-import { IconFace } from "~/client/components/BorderedIcon";
-import { BorderedIconRarity } from "~/client/components/BorderedIcon/BorderedIconRarity";
 import { useCCMap } from "~/client/contexts";
+import { IconFace } from "~/components/BorderedIcon";
+import { BorderedIconRarity } from "~/components/BorderedIcon/BorderedIconRarity";
 import { CCBackgroundFromBorder } from "~/items/ccBackgrounds";
 import type { ComponentPropsCC } from "~/types/ComponentProps";
 import styles from "./CommandCodeIcon.module.css";
