+++ src/client/components/EventList/EventList.tsx
@@ -1,12 +1,12 @@
 import Image from "next/image";
 import Link from "next/link";
-import Headline from "~/client/components/Headline";
-import { IconHourglass } from "~/client/components/icons";
-import { InlineIcon } from "~/client/components/InlineIcon";
-import Section from "~/client/components/Section";
-import { DisplayDate, DisplayDelta } from "~/client/components/TimeDisplay";
-import { useCurrentTime } from "~/client/utils/hooks/useCurrentTime";
-import { useIsClient } from "~/client/utils/hooks/useIsClient";
+import Headline from "~/components/Headline";
+import { IconHourglass } from "~/components/icons";
+import { InlineIcon } from "~/components/InlineIcon";
+import Section from "~/components/Section";
+import { DisplayDate, DisplayDelta } from "~/components/TimeDisplay";
+import { useCurrentTime } from "~/hooks/useCurrentTime";
+import { useIsClient } from "~/hooks/useIsClient";
 import { normalizeDate } from "~/time/normalizeDate";
 import type { BasicEvent } from "~/events/types";
 import styles from "./EventList.module.css";
