+++ src/client/components/NoSSR/NoSSR.tsx
@@ -1,4 +1,4 @@
-import { useIsClient } from "~/client/utils/hooks/useIsClient";
+import { useIsClient } from "~/hooks/useIsClient";
 
 export function NoSSR({ children }: React.PropsWithChildren) {
   const isClient = useIsClient();
