+++ src/pages/EventPage/components/EventInfoSection.tsx
@@ -1,12 +1,12 @@
 import spacetime from "spacetime";
-import { LinkButton } from "~/client/components/Button";
-import { DiscordTSButton } from "~/client/components/DiscordTSButton";
-import Headline from "~/client/components/Headline";
-import { Progress } from "~/client/components/Progress";
-import Section from "~/client/components/Section";
-import { DisplayDate } from "~/client/components/TimeDisplay";
-import { useCurrentTime } from "~/client/utils/hooks/useCurrentTime";
-import { useIsClient } from "~/client/utils/hooks/useIsClient";
+import { LinkButton } from "~/components/Button";
+import { DiscordTSButton } from "~/components/DiscordTSButton";
+import Headline from "~/components/Headline";
+import { Progress } from "~/components/Progress";
+import Section from "~/components/Section";
+import { DisplayDate } from "~/components/TimeDisplay";
+import { useCurrentTime } from "~/hooks/useCurrentTime";
+import { useIsClient } from "~/hooks/useIsClient";
 import { formatDiff } from "~/time/formatDiff";
 import { normalizeDate } from "~/time/normalizeDate";
 import type { BundledEvent } from "~/events/types";
