+++ src/pages/EventPage/components/EventListItem.tsx
@@ -3,11 +3,11 @@ import {
   BorderedItemIcon,
   BorderedServantIcon
 } from "~/client/components/BorderedIcon";
-import { TimerListEntities, TimerListItem } from "~/client/components/Card";
-import { DisplayDate, DisplayDelta } from "~/client/components/TimeDisplay";
-import { useCurrentTime } from "~/client/utils/hooks/useCurrentTime";
-import { useIsClient } from "~/client/utils/hooks/useIsClient";
+import { TimerListEntities, TimerListItem } from "~/components/Card";
 import { CommandCodeIcon } from "~/components/CommandCodeIcon";
+import { DisplayDate, DisplayDelta } from "~/components/TimeDisplay";
+import { useCurrentTime } from "~/hooks/useCurrentTime";
+import { useIsClient } from "~/hooks/useIsClient";
 import { normalizeDate } from "~/time/normalizeDate";
 import type { BundledEvent } from "~/events/types";
