+++ src/pages/EventPage/components/EventNewsModal.tsx
@@ -1,6 +1,6 @@
-import { ActionButton } from "~/client/components/Button";
-import { IconClose } from "~/client/components/icons";
-import { Modal } from "~/client/components/Modal";
+import { ActionButton } from "~/components/Button";
+import { IconClose } from "~/components/icons";
+import { Modal } from "~/components/Modal";
 import styles from "./EventNewsModal.module.css";
 
 interface EventNewsModalProps {
