+++ src/pages/UpgradesPage/components/FiltersForm.tsx
@@ -3,12 +3,12 @@ import cc from "classcat";
 import FlagEN from "flag-icons/flags/4x3/gb.svg";
 // @ts-ignore
 import FlagJP from "flag-icons/flags/4x3/jp.svg";
-import { ActionButton } from "~/client/components/Button";
-import { Pending } from "~/client/components/Pending";
-import { Selector } from "~/client/components/Selector";
+import { ActionButton } from "~/components/Button";
+import { Pending } from "~/components/Pending";
+import { Selector } from "~/components/Selector";
 import { getClassIconPath, nameServantClass } from "~/servants/classNames";
 import { UpgradeQuestType } from "~/upgrades/types";
-import type { SelectorOption } from "~/client/components/Selector";
+import type { SelectorOption } from "~/components/Selector";
 import { selectableClasses } from "../filtersReducer";
 import type { FormFilterAction, FormFilterState } from "../filtersReducer";
 import styles from "./FiltersForm.module.css";
